The simple part of the card have been designed with the responsive
The most difficult part to do was the responsive of the img, cause
we want to have an img that not just adapt with the property :
`max-width: 100%`, cause we lost some part of the img, this way, when
we screen size is getting bigger.

The key was to set the height of the img to 100%, and to use :

`object-fit: cover`. It helps us to save the img and not have dark
background, even if the size of the browser windows changes!